requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
coverage==7.6.9
dnspython==2.7.0
exceptiongroup==1.2.2
grpcio==1.68.1
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pymongo==4.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.0
requests==2.32.3
rollbar==1.1.0
ruff==0.8.3
tomli==2.2.1
urllib3==2.2.3
motor==3.3.0
python-dotenv==1.0.1
fastAPI==0.115.6
httpx==0.28.1
aiohttp==3.11.10
minio==7.2.12
python-multipart==0.0.19
uvicorn==0.34.0
